+++ Savage Hotel System/Savage Hotel System/Views/Reserva_List.cs
@@ -267,39 +267,46 @@ namespace Savage_Hotel_System.Views
 
         private void button2_Click(object sender, EventArgs e)
         {
-            DataGridViewCell selecionada = reservaDataGridView.SelectedCells[0];
-            if (selecionada != null)
+            if (changedCells.Count() == 0)
             {
-                string nomeLinha = reservaDataGridView["dataGridViewTextBoxColumn1", selecionada.RowIndex].Value.ToString();
-                DialogResult result1 = MessageBox.Show("Tem certeza que deseja cancelar a Reserva de ID: " + nomeLinha + " ?", "Remoção irreversível!", MessageBoxButtons.YesNo);
-
-                if (result1 == DialogResult.Yes)
+                DataGridViewCell selecionada = reservaDataGridView.SelectedCells[0];
+                if (selecionada != null)
                 {
+                    string nomeLinha = reservaDataGridView["dataGridViewTextBoxColumn1", selecionada.RowIndex].Value.ToString();
+                    DialogResult result1 = MessageBox.Show("Tem certeza que deseja cancelar a Reserva de ID: " + nomeLinha + " ?", "Remoção irreversível!", MessageBoxButtons.YesNo);
 
-                    //Id no banco da linha para excluir
-                    string idLinha = reservaDataGridView["dataGridViewTextBoxColumn1", selecionada.RowIndex].Value.ToString();
+                    if (result1 == DialogResult.Yes)
+                    {
 
-                    string queryString = "DELETE FROM " + DataBase.tableReserva + " Where Id = @idLinha";
-                    //chamando função da query paramateros (querystring, lista parametros, lista valores)
-                    var reader = DataBase.SqlCommand(queryString,
-                        new List<string>() {
+                        //Id no banco da linha para excluir
+                        string idLinha = reservaDataGridView["dataGridViewTextBoxColumn1", selecionada.RowIndex].Value.ToString();
+
+                        string queryString = "DELETE FROM " + DataBase.tableReserva + " Where Id = @idLinha";
+                        //chamando função da query paramateros (querystring, lista parametros, lista valores)
+                        var reader = DataBase.SqlCommand(queryString,
+                            new List<string>() {
                                  "@idLinha"
 
-                          }, new List<object>() {
+                              }, new List<object>() {
                                       idLinha
-                    });
+                        });
 
-                    //fechando a query, causa erros se nao fechar
-                    reader.Close();
-                    //remove da view
-                    reservaDataGridView.Rows.Remove(selecionada.OwningRow);
+                        //fechando a query, causa erros se nao fechar
+                        reader.Close();
+                        //remove da view
+                        reservaDataGridView.Rows.Remove(selecionada.OwningRow);
+
+                    }
 
                 }
-
-            }
-            else
+                else
+                {
+                    MessageBox.Show("Nenhuma Dado deletado", "Selecione pelo menos um campo para escluir a linha!", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+                }
+            }else
             {
-                MessageBox.Show("Nenhuma Dado deletado", "Selecione pelo menos um campo para escluir a linha!", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
+
+                MessageBox.Show("Confirme ou descarte as Alterações pendentes antes De cancelar uma Reserva!", "Alterações Pendentes", MessageBoxButtons.OK, MessageBoxIcon.Exclamation);
             }
         }
